+++ designer/server/src/main/resources/db/migration/postgres/V1_049__CreateFingerprintTable.sql
@@ -1,0 +1,3 @@
+CREATE TABLE "fingerprints" (
+   "value" VARCHAR(32) NOT NULL PRIMARY KEY
+);