+++ engine/development/deploymentManager/src/main/scala/pl/touk/nussknacker/development/manager/DevelopmentDeploymentManagerProvider.scala
@@ -85,32 +85,32 @@ class DevelopmentDeploymentManager(actorSystem: ActorSystem)
 
   }
 
-  override def processCommand[Result](command: ScenarioCommand[Result]): Future[Result] = command match {
-    case ValidateScenarioCommand(_, _, canonicalProcess) =>
+  override def processCommand[Result](command: DMScenarioCommand[Result]): Future[Result] = command match {
+    case DMValidateScenarioCommand(_, _, canonicalProcess) =>
       if (description(canonicalProcess).contains(descriptionForValidationFail)) {
         Future.failed(new IllegalArgumentException("Scenario validation failed as description contains 'fail'"))
       } else {
         Future.successful(())
       }
-    case command: RunDeploymentCommand                      => runDeployment(command)
-    case StopDeploymentCommand(name, _, savepointDir, user) =>
+    case command: DMRunDeploymentCommand                      => runDeployment(command)
+    case DMStopDeploymentCommand(name, _, savepointDir, user) =>
       // TODO: stopping specific deployment
-      stopScenario(StopScenarioCommand(name, savepointDir, user))
-    case command: StopScenarioCommand           => stopScenario(command)
-    case CancelDeploymentCommand(name, _, user) =>
+      stopScenario(DMStopScenarioCommand(name, savepointDir, user))
+    case command: DMStopScenarioCommand           => stopScenario(command)
+    case DMCancelDeploymentCommand(name, _, user) =>
       // TODO: cancelling specific deployment
-      cancelScenario(CancelScenarioCommand(name, user))
-    case command: CancelScenarioCommand  => cancelScenario(command)
-    case command: CustomActionCommand    => invokeCustomAction(command)
-    case _: MakeScenarioSavepointCommand => Future.successful(SavepointResult(""))
-    case _: TestScenarioCommand          => notImplemented
+      cancelScenario(DMCancelScenarioCommand(name, user))
+    case command: DMCancelScenarioCommand  => cancelScenario(command)
+    case command: DMCustomActionCommand    => invokeCustomAction(command)
+    case _: DMMakeScenarioSavepointCommand => Future.successful(SavepointResult(""))
+    case _: DMTestScenarioCommand          => notImplemented
   }
 
   private def description(canonicalProcess: CanonicalProcess) = {
     canonicalProcess.metaData.additionalFields.description
   }
 
-  private def runDeployment(command: RunDeploymentCommand): Future[Option[ExternalDeploymentId]] = {
+  private def runDeployment(command: DMRunDeploymentCommand): Future[Option[ExternalDeploymentId]] = {
     import command._
     logger.debug(s"Starting deploying scenario: ${processVersion.processName}..")
     val previous                = memory.get(processVersion.processName)
@@ -138,7 +138,7 @@ class DevelopmentDeploymentManager(actorSystem: ActorSystem)
     result.future
   }
 
-  private def stopScenario(command: StopScenarioCommand): Future[SavepointResult] = {
+  private def stopScenario(command: DMStopScenarioCommand): Future[SavepointResult] = {
     import command._
     logger.debug(s"Starting stopping scenario: $scenarioName..")
     asyncChangeState(scenarioName, Finished)
@@ -146,7 +146,7 @@ class DevelopmentDeploymentManager(actorSystem: ActorSystem)
     Future.successful(SavepointResult(""))
   }
 
-  private def cancelScenario(command: CancelScenarioCommand): Future[Unit] = {
+  private def cancelScenario(command: DMCancelScenarioCommand): Future[Unit] = {
     import command._
     logger.debug(s"Starting canceling scenario: $scenarioName..")
     changeState(scenarioName, DuringCancel)
@@ -166,7 +166,7 @@ class DevelopmentDeploymentManager(actorSystem: ActorSystem)
 
   override def customActionsDefinitions: List[CustomActionDefinition] = customActionStatusMapping.keys.toList
 
-  private def invokeCustomAction(command: CustomActionCommand): Future[CustomActionResult] = {
+  private def invokeCustomAction(command: DMCustomActionCommand): Future[CustomActionResult] = {
     val processName = command.processVersion.processName
     val statusOpt = customActionStatusMapping
       .collectFirst { case (customAction, status) if customAction.actionName == command.actionName => status }