+++ engine/flink/management/periodic/src/test/scala/pl/touk/nussknacker/engine/management/periodic/PeriodicDeploymentManagerTest.scala
@@ -59,10 +59,6 @@ class PeriodicDeploymentManagerTest
     val jarManagerStub                = new JarManagerStub
     val preparedDeploymentData        = DeploymentData.withDeploymentId(UUID.randomUUID().toString)
 
-    val deploymentService: ProcessingTypeDeploymentService = new ProcessingTypeDeploymentServiceStub(
-      List.empty
-    )
-
     val periodicProcessService = new PeriodicProcessService(
       delegateDeploymentManager = delegateDeploymentManagerStub,
       jarManager = jarManagerStub,
@@ -73,7 +69,7 @@ class PeriodicDeploymentManagerTest
       executionConfig = executionConfig,
       processConfigEnricher = ProcessConfigEnricher.identity,
       clock = Clock.systemDefaultZone(),
-      deploymentService
+      new ProcessingTypeActionServiceStub
     )
 
     val periodicDeploymentManager = new PeriodicDeploymentManager(
@@ -192,13 +188,13 @@ class PeriodicDeploymentManagerTest
     val emptyScenario = CanonicalProcess(MetaData("fooId", StreamMetaData()), List.empty)
 
     val validateResult = f.periodicDeploymentManager
-      .processCommand(ValidateScenarioCommand(processVersion, DeploymentData.empty, emptyScenario))
+      .processCommand(DMValidateScenarioCommand(processVersion, DeploymentData.empty, emptyScenario))
       .failed
       .futureValue
     validateResult shouldBe a[PeriodicProcessException]
 
     val deploymentResult = f.periodicDeploymentManager
-      .processCommand(RunDeploymentCommand(processVersion, DeploymentData.empty, emptyScenario, None))
+      .processCommand(DMRunDeploymentCommand(processVersion, DeploymentData.empty, emptyScenario, None))
       .failed
       .futureValue
     deploymentResult shouldBe a[PeriodicProcessException]
@@ -209,7 +205,12 @@ class PeriodicDeploymentManagerTest
 
     f.periodicDeploymentManager
       .processCommand(
-        RunDeploymentCommand(processVersion, f.preparedDeploymentData, PeriodicProcessGen.buildCanonicalProcess(), None)
+        DMRunDeploymentCommand(
+          processVersion,
+          f.preparedDeploymentData,
+          PeriodicProcessGen.buildCanonicalProcess(),
+          None
+        )
       )
       .futureValue
 
@@ -223,7 +224,7 @@ class PeriodicDeploymentManagerTest
 
     f.periodicDeploymentManager
       .processCommand(
-        RunDeploymentCommand(
+        DMRunDeploymentCommand(
           processVersion,
           DeploymentData.empty,
           PeriodicProcessGen.buildCanonicalProcess("0 0 0 ? * * 2000"),
@@ -243,7 +244,12 @@ class PeriodicDeploymentManagerTest
 
     f.periodicDeploymentManager
       .processCommand(
-        RunDeploymentCommand(processVersion, f.preparedDeploymentData, PeriodicProcessGen.buildCanonicalProcess(), None)
+        DMRunDeploymentCommand(
+          processVersion,
+          f.preparedDeploymentData,
+          PeriodicProcessGen.buildCanonicalProcess(),
+          None
+        )
       )
       .futureValue
 
@@ -273,7 +279,12 @@ class PeriodicDeploymentManagerTest
 
     f.periodicDeploymentManager
       .processCommand(
-        RunDeploymentCommand(processVersion, f.preparedDeploymentData, PeriodicProcessGen.buildCanonicalProcess(), None)
+        DMRunDeploymentCommand(
+          processVersion,
+          f.preparedDeploymentData,
+          PeriodicProcessGen.buildCanonicalProcess(),
+          None
+        )
       )
       .futureValue
 
@@ -294,7 +305,12 @@ class PeriodicDeploymentManagerTest
 
     f.periodicDeploymentManager
       .processCommand(
-        RunDeploymentCommand(processVersion, f.preparedDeploymentData, PeriodicProcessGen.buildCanonicalProcess(), None)
+        DMRunDeploymentCommand(
+          processVersion,
+          f.preparedDeploymentData,
+          PeriodicProcessGen.buildCanonicalProcess(),
+          None
+        )
       )
       .futureValue
 
@@ -316,7 +332,12 @@ class PeriodicDeploymentManagerTest
 
     f.periodicDeploymentManager
       .processCommand(
-        RunDeploymentCommand(processVersion, f.preparedDeploymentData, PeriodicProcessGen.buildCanonicalProcess(), None)
+        DMRunDeploymentCommand(
+          processVersion,
+          f.preparedDeploymentData,
+          PeriodicProcessGen.buildCanonicalProcess(),
+          None
+        )
       )
       .futureValue
 
@@ -338,7 +359,12 @@ class PeriodicDeploymentManagerTest
 
     f.periodicDeploymentManager
       .processCommand(
-        RunDeploymentCommand(processVersion, f.preparedDeploymentData, PeriodicProcessGen.buildCanonicalProcess(), None)
+        DMRunDeploymentCommand(
+          processVersion,
+          f.preparedDeploymentData,
+          PeriodicProcessGen.buildCanonicalProcess(),
+          None
+        )
       )
       .futureValue
 
@@ -361,7 +387,7 @@ class PeriodicDeploymentManagerTest
     f.repository.deploymentEntities.loneElement.status shouldBe PeriodicProcessDeploymentStatus.Failed
     f.repository.processEntities.loneElement.active shouldBe true
 
-    f.periodicDeploymentManager.processCommand(CancelScenarioCommand(processName, User("test", "Tester"))).futureValue
+    f.periodicDeploymentManager.processCommand(DMCancelScenarioCommand(processName, User("test", "Tester"))).futureValue
 
     f.repository.processEntities.loneElement.active shouldBe false
     f.repository.deploymentEntities.loneElement.status shouldBe PeriodicProcessDeploymentStatus.Failed
@@ -390,7 +416,7 @@ class PeriodicDeploymentManagerTest
     val deploymentId = f.repository.addActiveProcess(processName, PeriodicProcessDeploymentStatus.Deployed)
     f.delegateDeploymentManagerStub.setStateStatus(ProblemStateStatus.Failed, Some(deploymentId))
 
-    f.periodicDeploymentManager.processCommand(CancelScenarioCommand(processName, User("test", "Tester"))).futureValue
+    f.periodicDeploymentManager.processCommand(DMCancelScenarioCommand(processName, User("test", "Tester"))).futureValue
 
     f.repository.processEntities.loneElement.active shouldBe false
     f.repository.deploymentEntities.loneElement.status shouldBe PeriodicProcessDeploymentStatus.Failed
@@ -412,7 +438,7 @@ class PeriodicDeploymentManagerTest
     f.repository.deploymentEntities.loneElement.status shouldBe PeriodicProcessDeploymentStatus.Failed
     f.repository.processEntities.loneElement.active shouldBe true
 
-    f.periodicDeploymentManager.processCommand(CancelScenarioCommand(processName, User("test", "Tester"))).futureValue
+    f.periodicDeploymentManager.processCommand(DMCancelScenarioCommand(processName, User("test", "Tester"))).futureValue
 
     f.repository.processEntities.loneElement.active shouldBe false
     f.repository.deploymentEntities.loneElement.status shouldBe PeriodicProcessDeploymentStatus.Failed