+++ docs/Changelog.md
@@ -13,6 +13,8 @@
 * [#5831](https://github.com/TouK/nussknacker/pull/5831) Fragment input parameters presets support
 * [#5780](https://github.com/TouK/nussknacker/pull/5780) Fixed Scala case classes serialization when a class has additional fields in its body
 * [#5848](https://github.com/TouK/nussknacker/pull/5848) Change colors across an entire user interface
+* [#5853](https://github.com/TouK/nussknacker/pull/5853) Add processing mode information and filtering to the components table
+* [#5843](https://github.com/TouK/nussknacker/pull/5843) Added new endpoint to provide statistics URL
 
 1.14.0 (21 Mar 2024)
 -------------------------